+++ src/bin/chithi-run.rs
@@ -6,7 +6,7 @@ use std::io;
 fn main() -> io::Result<()> {
     let args = RunArgs::parse();
 
-    env_logger::Builder::from_env(env_logger::Env::default().default_filter_or("error"))
+    env_logger::Builder::from_env(env_logger::Env::default().default_filter_or("info"))
         .format_timestamp(None)
         .format_target(false)
         .init();